リファクタリング readme更新 パッケージ化

+++ ZIPFileControl/Assets/ZipFileControl/Scripts/ZipControl.cs
@@ -1,13 +1,13 @@
 ﻿using System;
 using System.Collections;
 using System.Collections.Generic;
-using System.Threading;
 using UnityEngine;
 using System.IO;
 #if UNITY_UWP
 using System.Threading.Tasks;
 using System.IO.Compression;
 #elif UNITY_EDITOR || UNITY_STANDALONE
+using System.Threading;
 using Ionic.Zip;
 #endif
 
@@ -15,28 +15,30 @@ namespace ZipFileControl
 {
     public class ZipControl
     {
-        public static IEnumerator OpenFile(string name,string outputpath,Action action=null)
+        /// <summary>
+        /// Zipファイルを展開
+        /// </summary>
+        /// <param name="name"></param>
+        /// <param name="filePath"></param>
+        /// <param name="action"></param>
+        /// <returns></returns>
+        public static IEnumerator OpenZip(string name,string filePath,Action action=null)
         {
 #if UNITY_UWP
-            Task task = Task.Run(() =>
+            var task = Task.Run(() =>
             {
                 if (File.Exists(name))
                 {
                     var directory = Path.GetDirectoryName(name);
-                    if (Directory.Exists(directory))
-                    {
-                        Directory.Delete(directory);
-                    }
+                    if (Directory.Exists(directory)) Directory.Delete(directory);
                     ZipFile.ExtractToDirectory(name, directory);
                 }
             });
             yield return new WaitWhile(() => task.IsCompleted == false);
 #elif UNITY_EDITOR || UNITY_STANDALONE
-            Thread thread = new Thread(()=> {
-                using (ZipFile zip=ZipFile.Read(name))
-                {
-                    zip.ExtractAll(outputpath);
-                }
+            var thread = new Thread(()=>
+            {
+                using (var zip = ZipFile.Read(name)) zip.ExtractAll(filePath);
             });
             thread.Start();
             yield return new WaitWhile(() => thread.IsAlive == true);
@@ -45,25 +47,28 @@ namespace ZipFileControl
             if (action != null) action.Invoke();
         }
 
-        public static IEnumerator CreateFile(string directory,string outputfile,Action action)
+        /// <summary>
+        /// Zipファイルを作成
+        /// </summary>
+        /// <param name="directory"></param>
+        /// <param name="filePath"></param>
+        /// <param name="action"></param>
+        /// <returns></returns>
+        public static IEnumerator CreateZip(string directory,string filePath,Action action)
         {
 #if UNITY_UWP
-            Task task = Task.Run(() =>
+            var task = Task.Run(() =>
             {
-                if (File.Exists(outputfile))
-                {
-                    File.Delete(outputfile);
-                }
-                ZipFile.CreateFromDirectory(directory, outputfile, CompressionLevel.NoCompression, true);
+                if (File.Exists(filePath)) File.Delete(filePath);
+                ZipFile.CreateFromDirectory(directory, filePath, CompressionLevel.NoCompression, true);
             });
             yield return new WaitWhile(() => task.IsCompleted == false);
 #elif UNITY_EDITOR || UNITY_STANDALONE
-            Thread thread = new Thread(() => {
-                using (ZipFile zip=new ZipFile())
+            var thread = new Thread(() => {
+                using (var zip=new ZipFile())
                 {
-                    var path = Path.GetFileName(directory);
-                    zip.AddDirectory(directory, path);
-                    zip.Save(outputfile);
+                    zip.AddDirectory(directory, Path.GetFileName(directory));
+                    zip.Save(filePath);
                 }
             });
             thread.Start();
